minor refactor on database deserialization
